Created module for chamford algorithm operations Also did some refactorization
