+++ segmentation.py
@@ -156,165 +156,3 @@ def determine_binary_color(binary: np.ndarray, x: int, y: int, r_original: float
         return 255 if white_count > black_count else 0
     else:  # bad input
         return -1
-
-
-def negativ(img: np.ndarray) -> np.ndarray:
-    """
-    :param img: input image
-    :return: negative image
-    """
-    width, height = img.shape[:2]
-    negativ_image = np.zeros((width, height), np.uint8)
-    for i in range(0, width):
-        for j in range(0, height):
-            negativ_image[i][j] = 255 - img[i][j]
-    return negativ_image
-
-
-def array_scope_control(img: np.ndarray, i: int, j: int) -> bool:
-    """
-    checks if (i,j) is in the table
-    :param img: input image
-    :param i:
-    :param j:
-    :return:
-    """
-    width, height = img.shape[:2]
-    if 0 <= i < width and 0 <= j < height:
-        return True
-    else:
-        return False
-
-
-def pixel_value(img: np.ndarray, i: int, j: int, direction: int) -> int:
-    """
-    Calculates the pixel(i,j) value for the chamford function
-    Forward: use    [*, *, *] mask
-                    [*, 0, -]
-                    [-, -, -]
-
-    Backward: use   [-, -, -] mask
-                    [-, 0, *]
-                    [*, *, *]
-
-    where '*' value is uses, '0' is center of mask (the center is located in pixel(i,j)), '-' value is non used
-
-    :param img: input image
-    :param i:
-    :param j:
-    :param direction: determines the phase forward or backward
-    :return: minimum pixel distance from the background
-    """
-    mask = [
-        [1, 1, 1],
-        [1, 0, 1],
-        [1, 1, 1]
-    ]  # determines the pixel distance from the center
-    minimum = []
-    if direction == 1:  # forward
-        i -= 1
-        j -= 1
-        for x in range(0, 3):
-            for y in range(0, 3):
-                if x == 1 and y == 1:
-                    return min(minimum)
-                if array_scope_control(img, i + x, j + y):
-                    if (img[i + x][j + y] + mask[x][y]) == 256:
-                        minimum.append(0)
-                    else:
-                        minimum.append(img[i + x][j + y] + mask[x][y])
-    else:  # backward
-        i += 1
-        j += 1
-        for x in range(0, 3):
-            for y in range(0, 3):
-                if x == 1 and y == 1:
-                    return min(minimum)
-                if array_scope_control(img, i - x, j - y):
-                    if (img[i - x][j - y] + mask[x][y]) == 256:
-                        minimum.append(0)
-                    else:
-                        minimum.append(img[i - x][j - y] + mask[x][y])
-
-
-def chamford(img: np.ndarray) -> np.ndarray:
-    """
-    The function implements a chamford algorithm that calculates the distance of the pixel from the background
-
-    Chamford algorithm have two steps:
-    Forward: goes right and down
-    Backward: goes left up
-
-    :param img: binary image. 255 pixel is background
-    :return: image with the distance of the image pixels from the background
-    """
-    width, height = img.shape[:2]
-    size = 3  # size mask in pixel_value
-    chamford_image = np.zeros((width, height), np.uint8)
-    # forward
-    for i in range(math.floor((size + 1) / 2), width):
-        for j in range(math.floor((size + 1) / 2), height):
-            if img[i][j] < 255:
-                chamford_image[i][j] = pixel_value(chamford_image, i, j, 1)
-            else:
-                chamford_image[i][j] = img[i][j]
-    # backward
-    for i in range(math.floor(width - (size - 1) / 2), 0, -1):
-        for j in range(math.floor(height - (size - 1) / 2), 0, -1):
-            if img[i][j] < 255:
-                val = pixel_value(chamford_image, i, j, 2)
-                if val < chamford_image[i][j]:
-                    chamford_image[i][j] = val
-
-    return chamford_image
-
-
-def spr_local_extreme(img: np.ndarray, i: int, j: int, val: int) -> bool:
-    """
-    The function determines whether pixel (i, j) is a local extreme
-
-    :param img: Input image
-    :param i:
-    :param j:
-    :param val: pixel distance from the background
-    :return: True if pixel[i][j] is local extreme, otherwaise False
-    """
-    i -= 1
-    j -= 1
-    for x in range(0, 3):
-        for y in range(0, 3):
-            if x == 1 and y == 1:
-                continue
-            if array_scope_control(img, i - x, j - y):
-                if img[i - x][j - y] > val and img[i - x][j - y] != 255:  # if the pixel is 255, its a background
-                    return False
-    return True
-
-
-def extreme(img: np.ndarray) -> (int,int) :
-    """
-    Image must be after chamford function, background pixel value must be 255
-    Finds min and max object extremum in image
-    Object extremum is the furthest pixel distance from the background
-
-
-    :param img: Image
-    :return: minimum,maximum
-    """
-    width, height = img.shape[:2]
-    extremes = []
-    max_extreme = 0
-    # find max extreme value
-    for i in range(0, width):
-        for j in range(0, height):
-            if max_extreme < img[i][j] < 255:
-                max_extreme = img[i][j]
-    # find all local extreme value
-    for val in range(1, max_extreme):
-        for i in range(0, width):
-            for j in range(0, height):
-                if img[i][j] == val:
-                    if spr_local_extreme(img, i, j, val):
-                        extremes.append(val)
-
-    return min(extremes), max_extreme
